requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
future==0.18.2
idna==2.10
importlib-metadata==4.0.1
joblib==1.0.1
krwordrank==1.0.3
Markdown==3.3.4
mysqlclient==2.0.3
normalize==2.0.2
numpy==1.21.1
Pillow==8.2.0
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
richenum==1.3.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.26.5
zipp==3.4.1
